third_party/WebKit/LayoutTests/svg/animations/svgpath-animation-1-expected.txt

SVG 1.1 dynamic animation tests
Test 'by' animation on path. You should see a green 100x100 path and only PASS messages

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS path.getAttribute('d') is "M 40 40 L 60 40 L 60 60 L 40 60 z"
PASS path.getAttribute('d') is "M 40 40 L 109.975 40 L 109.975 109.975 L 40 109.975 Z"
PASS path.getAttribute('d') is "M 40 40 L 110.025 40 L 110.025 110.025 L 40 110.025 Z"
PASS path.getAttribute('d') is "M 40 40 L 159.975 40 L 159.975 159.975 L 40 159.975 Z"
PASS path.getAttribute('d') is "M 40 40 L 160 40 L 160 160 L 40 160 Z"
PASS successfullyParsed is true

TEST COMPLETE
